# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Gin Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/cart": {
            "get": {
                "description": "Get Items From Cart.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Get Items From Cart.",
                "parameters": [
                    {
                        "description": "Cart JSON",
                        "name": "cart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.UserCart"
                        }
                    }
                }
            },
            "post": {
                "description": "Add Item To Cart.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Add Item To Cart.",
                "parameters": [
                    {
                        "description": "Cart JSON",
                        "name": "cart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ItemToCartUserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Remove Item From Cart.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Remove Item From Cart.",
                "parameters": [
                    {
                        "description": "Cart JSON",
                        "name": "cart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ItemToCartUserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order": {
            "post": {
                "description": "Create order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Create order.",
                "parameters": [
                    {
                        "description": "Create Order JSON",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order-status": {
            "post": {
                "description": "Changed status to order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Changed status to order.",
                "parameters": [
                    {
                        "description": "Order Change Status JSON",
                        "name": "OrderStatus",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateOrderStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/orders": {
            "post": {
                "description": "Get orders.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Get orders.",
                "parameters": [
                    {
                        "description": "Get Orders JSON",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserOrder"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ItemDao": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pricePerUnit": {
                    "type": "integer"
                },
                "priceTotal": {
                    "type": "integer"
                }
            }
        },
        "models.ItemDto": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "itemName": {
                    "type": "string"
                }
            }
        },
        "models.ItemToCartUserDto": {
            "type": "object",
            "properties": {
                "item": {
                    "$ref": "#/definitions/models.ItemDto"
                },
                "user": {
                    "$ref": "#/definitions/models.UserDto"
                }
            }
        },
        "models.OrderStatus": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.UpdateOrderStatusDto": {
            "type": "object",
            "properties": {
                "statusName": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "models.UserCart": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ItemDao"
                    }
                },
                "user": {
                    "$ref": "#/definitions/models.UserDto"
                }
            }
        },
        "models.UserDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "models.UserOrder": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ItemDao"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.OrderStatus"
                },
                "user": {
                    "$ref": "#/definitions/models.UserDto"
                }
            }
        }
    }
}